+++ src/main/java/dev/enjarai/trickster/block/SpellConstructBlock.java
@@ -33,7 +33,7 @@ import org.jetbrains.annotations.Nullable;
 
 public class SpellConstructBlock extends BlockWithEntity {
     public static final DirectionProperty FACING = Properties.FACING;
-    public static final VoxelShape[] SHAPES = new VoxelShape[]{
+    public static final VoxelShape[] SHAPES = new VoxelShape[] {
             createCuboidShape(0, 14, 0, 16, 16, 16),
             createCuboidShape(0, 0, 0, 16, 2, 16),
             createCuboidShape(0, 0, 14, 16, 16, 16),
@@ -45,7 +45,8 @@ public class SpellConstructBlock extends BlockWithEntity {
     protected SpellConstructBlock() {
         super(AbstractBlock.Settings.create()
                 .strength(1.5F)
-                .sounds(BlockSoundGroup.STONE));
+                .sounds(BlockSoundGroup.STONE)
+                .solid());
         setDefaultState(stateManager.getDefaultState().with(FACING, Direction.UP));
     }